Create Get - Software Uninstall String from Registry.ps1
Simple script to find the uninstall string for an application in the Windows Registry without the use of Win32_Product which is problematic. Uses Read-Host to ask the console user for the Application Name. Outputs Application DisplayName and both the UninstallString and QuietUninstallString registry entries. Works well with Enter-PSsession and aids in creating Application Models in SCCM/MECM.